DI tidy up for autofill
