+++ app/src/main/res/raw/inject_alias.js
