+++ modules/autofill/autofill-impl/build.gradle
@@ -31,6 +31,7 @@ dependencies {
     implementation project(path: ':autofill-api')
     implementation project(path: ':secure-storage-api')
     implementation project(path: ':device-auth-api')
+    implementation project(path: ':browser-api')
     api project(path: ':autofill-store')
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
